Add Content, Font Awesome
Added Font Awesome for brand icons on the about me page as part of migrating
from my old personal site.

Started the layout of the blog with the first entry, and included pages
for common resources I recommend to folks interested in coding.

This did NOT include formatting as that should be the first commit before
then creating new resources, instead it will come after this.